requirements.txt
ollama==0.1.5
strands
tweepy
python-dotenv
